LayoutTests/http/tests/security/mixedContent/insecure-xhr-in-main-frame-expected.txt

CONSOLE MESSAGE: [blocked] The page at https://127.0.0.1:8443/security/mixedContent/resources/insecure-xhr-in-main-frame-window.html requested insecure content from http://127.0.0.1:8000/. This content was blocked and must be served over HTTPS.

CONSOLE MESSAGE: Not allowed to request resource
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8000/ due to access control checks.
This test opens a HTTPS window that loads insecure data via XHR. We should trigger a mixed content callback because the main frame in the window is HTTPS but now has insecure data.
